moves interface checker to checker folder
