+++ backend/typescript/middlewares/validators/reviewValidators.ts
@@ -1,6 +1,6 @@
 import { Request, Response, NextFunction } from "express";
 import { createCheckers } from "ts-interface-checker";
-import IReviewTI from "../../services/interfaces/IReviewService-ti";
+import IReviewTI from "../../services/interfaces/checkers/IReviewService-ti";
 import { getErrorMessage } from "../../utilities/errorResponse";
 
 const reviewRequestDtoValidator = async (
